File rename to reflect that backend is based on dash_map
