+++ memcrs/src/memcache/builder.rs
@@ -1,6 +1,6 @@
 use crate::cache::eviction_policy::EvictionPolicy;
 use crate::cache::cache::Cache;
-use crate::memory_store::store::{MemoryStore as DashMapStore};
+use crate::memory_store::dash_map_store::{MemoryStore as DashMapStore};
 use crate::memory_store::moka_store::{MemoryStore as MokaStore};
 use crate::memory_store::StoreEngine;
 use crate::server::timer;
