+++ memcrs/src/mock/mock_server.rs
@@ -1,7 +1,7 @@
 use crate::memcache::store::MemcStore;
 
 use crate::memory_store::moka_store::MemoryStore as MokaStore;
-use crate::memory_store::store::MemoryStore;
+use crate::memory_store::dash_map_store::MemoryStore;
 use crate::server::timer;
 use std::sync::atomic::{AtomicUsize, Ordering};
 use std::sync::Arc;
